input refactor and improvement, can now have multiple inputs (both gamepad and keybaord), or (keyboard & keyyboard) and so on, all of these work similatanously, though support for more buttons and stuff is neded when it comes to gamepad input.

+++ source/Engine/Pinecorn/Input/VirtualAxisKeyboard.cs
@@ -11,7 +11,7 @@ namespace Pinecorn
             Negative = negative;
         }
 
-        public override int GetAxis()
+        public override float GetAxis()
         {
             if (Positive.Pressed())
                 return 1;
